pyproject.toml
[tool.ruff]
ignore = ["F403", "E402", "F405"]
target-version = "py312"
# Same as Black.
line-length = 120
indent-width = 4

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
# Like Black, indent with spaces, rather than tabs.
indent-style = "space"
# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false
# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.pdm]
distribution = false

[project]
name = "ME_Blender_Toolbox"
version = "0.1.0"
description = "Metric Empire Blender Addon"
authors = [{ name = "Metric Empire", email = "hello@metricempire.com" }]
dependencies = [
    "toml>=0.10.2",
    "fake-bpy-module-latest>=20241217",
    "pyme @ git+https://metricempire@dev.azure.com/metricempire/MEDevOps/_git/pyme@main"
]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "proprietary" }

[tool.ruff.lint]
extend-ignore = []
extend-select = []
fixable = ["ALL"]
